Added changes from pull request
