--- modules/iam-types/src/main/scala/ch/epfl/bluebrain/nexus/commons/iam/types/acls/AccessControl.scala
@@ -1,8 +1,0 @@
-package ch.epfl.bluebrain.nexus.commons.iam.types.acls
-
-import ch.epfl.bluebrain.nexus.commons.iam.types.identity.Identity
-
-/**
-  * Type definition that is essentially a pair consisting of an ''identity'' and its associated ''permissions''.
-  */
-final case class AccessControl(identity: Identity, permissions: Permissions)
