--- modules/iam-types/src/test/scala/ch/epfl/bluebrain/nexus/commons/iam/types/acls/PermissionSpec.scala
@@ -1,14 +1,0 @@
-package ch.epfl.bluebrain.nexus.commons.iam.types.acls
-
-import org.scalatest.{Matchers, WordSpecLike}
-
-class PermissionSpec extends WordSpecLike with Matchers {
-  "A Permission" should {
-
-    "should match the regex" in {
-      intercept[IllegalArgumentException] {
-        Permission("One")
-      }
-    }
-  }
-}
