--- modules/iam-types/src/test/scala/ch/epfl/bluebrain/nexus/commons/iam/types/acls/PermissionsSpec.scala
@@ -1,42 +1,0 @@
-package ch.epfl.bluebrain.nexus.commons.iam.types.acls
-
-import ch.epfl.bluebrain.nexus.commons.iam.types.acls.Permission._
-
-import org.scalatest.{Matchers, WordSpecLike}
-
-class PermissionsSpec extends WordSpecLike with Matchers {
-  "A Permissions" should {
-    val permissions = Permissions(Write)
-
-    "add a permission to the permissions" in {
-      permissions + Read shouldEqual Permissions(Write, Read)
-    }
-
-    "subtract a permission to the permissions" in {
-      (permissions + Read - Write) shouldEqual Permissions(Read)
-    }
-
-    "add permissions to the permissions" in {
-      permissions ++ Permissions(Read, Own) shouldEqual Permissions(Write, Read, Own)
-    }
-
-    "subtract permissions to the permissions" in {
-      (permissions + Read) -- Permissions(Write, Own) shouldEqual Permissions(Read)
-    }
-
-    "show permissions with match two permissions" in {
-      Permissions(Read, Write, Own) intersect (Permissions(Write)) shouldEqual (Permissions(Write))
-      Permissions(Read, Write) & (Permissions(Own)) shouldEqual (Permissions())
-    }
-
-    "check if a permission is contained on permissions" in {
-      Permissions(Read, Write).contains(Read) shouldEqual true
-      Permissions(Read, Write).contains(Own) shouldEqual false
-    }
-
-    "check if some permissions are contained on permissions" in {
-      Permissions(Read, Write).containsAll(Permissions(Read)) shouldEqual true
-      Permissions(Read, Write).containsAll(Permissions(Read, Own)) shouldEqual false
-    }
-  }
-}
