+++ modules/iam/src/test/scala/ch/epfl/bluebrain/nexus/commons/iam/acls/PermissionSpec.scala
@@ -1,0 +1,14 @@
+package ch.epfl.bluebrain.nexus.commons.iam.acls
+
+import org.scalatest.{Matchers, WordSpecLike}
+
+class PermissionSpec extends WordSpecLike with Matchers {
+  "A Permission" should {
+
+    "should match the regex" in {
+      intercept[IllegalArgumentException] {
+        Permission("One")
+      }
+    }
+  }
+}
